Try to add delete product file

+++ products/views.py
@@ -132,6 +132,19 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+# @login_required
+# def delete_product(request, product_id):
+#     """ Delete a product from the store """
+#     if not request.user.is_superuser:
+#         messages.error(request, 'Sorry, only store owners can do that.')
+#         return redirect(reverse('home'))
+
+#     product = get_object_or_404(Product, pk=product_id)
+#     product.delete()
+#     messages.success(request, 'Product deleted!')
+#     return redirect(reverse('products'))
+
+
 @login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
@@ -139,7 +152,11 @@ def delete_product(request, product_id):
         messages.error(request, 'Sorry, only store owners can do that.')
         return redirect(reverse('home'))
 
-    product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+    if request.method == 'POST':
+        product = get_object_or_404(Product, pk=product_id)
+        product.delete()
+        messages.success(request, 'Product deleted!')
+        return redirect(reverse('products'))
+    else:
+        # Handle GET request or any other condition
+        return redirect(reverse('products'))
